requirements.txt
python==3.10.4
qiskit==0.37.1
qiskit_optimization==0.4.0
matplotlib==3.5.2
pyyaml==6.0
pandas==1.4.3
ROOT==6.26/00
h5py==3.7.0